// Finanzuebersicht.Frontend.Web.Admin/src/app/model/accounting/accounting-entries/dtos/i-accounting-entry-detail.ts
import { ICategory } from 'src/app/model/accounting/categories/dtos/i-category';

export interface IAccountingEntryDetail {
    id: string;
    category: ICategory;
    auftragskonto: string;
    buchungsdatum: Date;
    valutaDatum: Date;
    buchungstext: string;
    verwendungszweck: string;
    glaeubigerId: string;
    mandatsreferenz: string;
    sammlerreferenz: string;
    lastschriftUrsprungsbetrag: number;
    auslagenersatzRuecklastschrift: string;
    beguenstigter: string;
    iban: string;
    bic: string;
    betrag: number;
    waehrung: string;
    info: string;
}